Se elimina pantalla duplicada de confirmarQuiz
